refactor(core): camera: move session callback out of session controller
